pyproject.toml
[project]
name = "anm-hass-integration"
version = "1.0.0"
description = "Home Assistant integration for ANM (Azienda Napoletana Mobilita) public transport"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "DLubrano", email = "your.email@example.com"}
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "aioresponses>=0.7.0",
    "coverage>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "C",
]
ignore = [
    "E501",  # Line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["custom_components"]
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
